revise folder structure, revise update profile
