+++ src/components/auth/UpdateProfile/UpdateProfile.js
@@ -1,25 +1,21 @@
-import React, { useRef, useState } from "react";
-import { useAuth } from "../../contexts/AuthContext"
-import { useModal } from "../../contexts/ModalContext"
+import React, { useRef, useState } from "react"
 import { useHistory, Link } from "react-router-dom";
-import ErrorModal from '../ErrorModal/ErrorModal'
-import './Signup.css'
+import { useModal } from "../../../contexts/ModalContext"
+import { useAuth } from "../../../contexts/AuthContext"
+import ErrorModal from '../../ErrorModal/ErrorModal'
 
-const Signup = () => {
-  // const displaynameRef = useRef()
+export default function UpdateProfile() {
   const emailRef = useRef()
   const passwordRef = useRef()
   const passwordConfirmRef = useRef()
-
-  const { signup } = useAuth()
+  const { currentUser, updatePassword, updateEmail } = useAuth()
   const { openErrorModal, errorMessageHandler } = useModal()
   const [error, setError] = useState("")
   const [loading, setLoading] = useState(false)
   const history = useHistory()
 
-  async function handleSubmit(e) {
+  function handleSubmit(e) {
     e.preventDefault()
-
     if (passwordRef.current.value !== passwordConfirmRef.current.value) {
       let message = "Passwords do not match"
       errorMessageHandler(message)
@@ -27,20 +23,33 @@ const Signup = () => {
       return setError(message)
     }
 
-    // the loading state is set to disable the button so only 1 account is created
-    try {
-      setError("")
-      setLoading(true)
-      await signup(emailRef.current.value, passwordRef.current.value)
-      history.push("/")
-    } catch {
-      let message = "Failed to create an account"
-      setError(message)
-      errorMessageHandler(message)
-      openErrorModal(true)
+    const promises = []
+    setLoading(true)
+    setError("")
+
+    console.log('Emails', emailRef.current.value, currentUser.email)
+    if (emailRef.current.value && emailRef.current.value !== currentUser.email) {
+      console.log('Updating email')
+      promises.push(updateEmail(emailRef.current.value))
+    }
+    if (passwordRef.current.value) {
+      promises.push(updatePassword(passwordRef.current.value))
     }
 
-    setLoading(false)
+    Promise.all(promises)
+      .then(() => {
+        history.push("/profile")
+      })
+      .catch((err) => {
+        let message = "Failed to update account"
+        console.log(err.message)
+        setError(message)
+        errorMessageHandler(message)
+        openErrorModal(true)
+      })
+      .finally(() => {
+        setLoading(false)
+      })
   }
 
   return (
@@ -49,7 +58,7 @@ const Signup = () => {
         <div className='form-wrapper'>
           {error && <ErrorModal />}
           <div className='login-form-outline'>
-            <h1>Sign Up</h1>
+            <h1>Update Profile</h1>
             <form onSubmit={handleSubmit}>
               {/* <div className="display-name">
               <label htmlFor="displayName">Display Name</label>
@@ -65,20 +74,18 @@ const Signup = () => {
                 <label htmlFor="email">Email</label>
                 <input
                   ref={emailRef}
-                  placeholder="Email"
+                  placeholder={currentUser.email}
                   type="email"
                   name="email"
-                  required
                 />
               </div>
               <div className="password">
                 <label htmlFor="password">Password</label>
                 <input
                   ref={passwordRef}
-                  placeholder="Password"
+                  placeholder="leave blank to keep the same"
                   type="password"
                   name="password"
-                  required
                   pattern="(?=.*\d)(?=.*[!@_#$%^&*-])(?=.*[a-z])(?=.*[A-Z]).{6,}"
                 />
               </div>
@@ -86,25 +93,22 @@ const Signup = () => {
                 <label htmlFor="password">Confirm Password</label>
                 <input
                   ref={passwordConfirmRef}
-                  placeholder="Password"
+                  placeholder="leave blank to keep the same"
                   type="password"
                   name="password"
-                  required
                   pattern="(?=.*\d)(?=.*[!@_#$%^&*-])(?=.*[a-z])(?=.*[A-Z]).{6,}"
                 />
               </div>
               <div className="createAccount">
-                <button disabled={loading} type="submit">Sign Up</button>
+                <button disabled={loading} type="submit">Update</button>
               </div>
             </form>
           </div>
           <div className="signup-login">
-            Already have an account? <Link to="/login">Log In</Link>
+            <Link to="/profile">Cancel</Link>
           </div>
         </div>
       </div>
     </div>
-  );
+  )
 }
-
-export default Signup;
